requirements.txt
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
ForgeryPy==0.1
html5lib==1.0.1
Werkzeug==0.14.1
WTForms==2.1
